Rework and finalize configuration:
The LoginPath that can be configured is now nullable and enable the handling of Challenge like, for example for WFA. It works also for any scheme that would not support Challenge. The only thing to configure is the login path.

There are also 2 ways to use the Asp package:
- AddOpenIddictAspWebFrontAuth => Out of the box solution. Still configurable.
- Combo of UseOpenIddictServerAsp and UseOpenIddictCoreSql.

UseOpenIddictCoreSql can be used alone of course when installing only the Sql package.