Add a strategy to validate the user and set the claims:
Each endpoint in the oidc process will call the 2 methods of the IIdentityStrategy.
This way, it is possible to support any kind of authentication method, either standard dotnet like Cookie, or WebFrontAuth.
WebFrontAuth got an example implementation in the sample.
First it validate that the user exists in the database, and can return the AuthenticationInfo.
Then it can set the identity claim depending on the existence of a scope. This scenario can be used depending on the client, that could have a storage or not.

AuthenticationInfo has 2 values that are not customizable because used in the controller: userId and userName.