Implement consent form:
The controller can redirect to the consent form url (hardcoded for now, will be parameterized) with all parameters in the query string.
The front end consent form can be 'accept' or 'deny'. It will pass all from the query as a form data.
The front end also sent the antiforgery header, that its get from the cookie.
About AntiForgery, the backend use a double cookie submit. The cookies are created only when a user is authenticated as we don't need it for login.