Provide only one way to configure the startup, clear and explicit:
The Sample still shows a WebFrontAuth configuration.
There is no weird development configuration that appear alone in the readme.
The developper can now copy paste the one and only way to create the startup, with minimal effort to change any value, while being able to understand directly what he does.
AddRouting is mandatory hence is registered with the Asp part.