Cris results are now bound to IPoco:
On the sample, the cksetup now generate typescript from Cris commands.
IApplicationPoco contains everything to describe an application to CRUD.
If it has an ApplicationId, it is considered as an existing application.
An OpenIddictApplicationDescriptor is used to create an application.
An Application can be used for all the rest.
In order to make the application manipulation as deep as possible from the frontend, the IApplicationPoco contains the maximum properties described as possible.