Add more way to update an application:
The existing one is renamed hard because it will override any property. Meaning that a null property will clear the cell in DB.
The soft one ignore null values. Collections are still overriden.
It is possible to rename quickly an application. This is just an example, any property could have its own command.