Handle ClientSecret:
When calling ApplicationManager.Update() there are 2 overloads to consider.
First, call with TApplication will perform a classic update field by field.
Second, call with TApplication and string secret will hash the secret and then update all the fields, with the hashed secret.

On both update methods, we ensure that the intention of the user is to remove the value (only for hard update), keep the value or change it.
Note that if a non hashed value of the actual ClientSecret is passed, it will be hashed again and the same ClientSecret will be used anyway.